pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lattice"
version = "0.1.0"
description = "A graph-based system modeling tool for capturing and validating system intent"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Lattice Team"}
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "networkx>=3.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "anthropic>=0.40.0",
]
semantic = [
    "anthropic>=0.40.0",
]

[project.scripts]
intent = "lattice.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["lattice*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
